pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.13"
description = ""
authors = ["kukhmax <kukhmax@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
types-PyYAML = "^5.4.10"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
autopep8 = "^1.5.7"
mypy = "^0.910"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
